requirements.txt
streamlit>=1.35.0
pandas>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.20.0
pgeocode>=0.5.0
